+++ image/install.sh
@@ -12,6 +12,7 @@ if [[ "$ruby30" = 1 ]]; then run /pd_build/ruby-3.0.*.sh; fi
 if [[ "$ruby31" = 1 ]]; then run /pd_build/ruby-3.1.*.sh; fi
 if [[ "$ruby32" = 1 ]]; then run /pd_build/ruby-3.2.*.sh; fi
 if [[ "$jruby93" = 1 ]]; then run /pd_build/jruby-9.3.*.sh; fi
+if [[ "$jruby94" = 1 ]]; then run /pd_build/jruby-9.4.*.sh; fi
 if [[ "$python" = 1 ]]; then run /pd_build/python.sh; fi
 if [[ "$nodejs" = 1 ]]; then run /pd_build/nodejs.sh; fi
 if [[ "$redis" = 1 ]]; then run /pd_build/redis.sh; fi