+++ image/utilities.sh
@@ -6,3 +6,5 @@ source /pd_build/buildconfig
 run minimal_apt_get_install build-essential
 ## Bundler has to be able to pull dependencies from git.
 run minimal_apt_get_install git
+## JRuby94 at least requires netbase, and other client stuff must.
+run minimal_apt_get_install netbase